feat(pool_budget_service): add pool budget calculations for regions
